Initial revamp of Physical oM
