+++ Physical_oM/Constructions/Layer.cs
@@ -27,16 +27,19 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
+using BH.oM.Geometry;
+using BH.oM.Physical.Materials;
 
-namespace BH.oM.Physical.Properties
+namespace BH.oM.Physical.Constructions
 {
-    public interface IFramingElementProperty : IBHoMObject
+    public class Layer : BHoMObject, IPhysical
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-
+        public Material Material { get; set; } = new Material();
+        public double Thickness { get; set; } = 0;
 
         /***************************************************/
     }
