+++ Physical_oM/Elements/Cable.cs
@@ -26,7 +26,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-using BH.oM.Physical.Properties;
 using BH.oM.Base;
 using BH.oM.Geometry;
